Close #38
H5CgnsFlowSolution::readValueNames() returns "GridLocation" as name of flow solution, but it actually is not a name of flow solution.

Currently, readValueNames() returns the names of all child groups in FlowSolution node. So, for FlowCellSolutionX, FlowIFaceSolutionX etc., it returns "GridLocation" as an element of the vector.
But it actually is not a name of flow solution, and that causes problems.

Fix the function so that it returns the names of groups with label "DataArray_t", that are actually groups that stores flow solutions.